+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 12.2.0 - 2023-11-02
+
+- Added `models get-data-bundle`
+- Added `models get-training`
+
 ## Version 12.1.0 - 2023-11-01
 
 - Updated `--create-default-workflow` in `workflows create-default` to use the latest docker images and automatically